+++ examples/demo.rs
@@ -98,7 +98,7 @@ async fn main() {
     };
 
     let topic =
-        unsafe { TopicName::new_unchecked(MqttString::from_slice("rust-mqtt/is/great").unwrap()) };
+        TopicName::new_checked(MqttString::from_slice("rust-mqtt/is/great").unwrap()).unwrap();
 
     match client.subscribe(topic.clone().into(), sub_options).await {
         Ok(_) => info!("Sent Subscribe"),